requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
happybase>=1.2.0
python-dotenv>=1.0.0
thrift>=0.16.0
